tg/requirements.txt
requests
newsapi-python
python-telegram-bot